+++ src/main/java/subway/dao/LineDao.java
@@ -5,7 +5,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.LineEntity;
+import subway.persistence.entity.LineEntity;
 
 import javax.sql.DataSource;
 import java.util.HashMap;
