+++ src/main/java/subway/dao/SectionDao.java
@@ -6,8 +6,8 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.SectionEntity;
-import subway.entity.SectionWithStationNameEntity;
+import subway.persistence.entity.SectionEntity;
+import subway.persistence.entity.SectionWithStationNameEntity;
 
 import javax.sql.DataSource;
 import java.util.List;
