+++ src/main/java/subway/dao/StationDao.java
@@ -7,7 +7,7 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import subway.entity.StationEntity;
+import subway.persistence.entity.StationEntity;
 
 import javax.sql.DataSource;
 import java.util.List;
